pyproject.toml
[tool.poetry]
name = "openai-autogen-dev-studio"
version = "0.1.0"
description = "Generate games and programs using OpenAI agents. Built on top of Microsoft Autogen."
authors = ["Ivan Gabriele <ivan.gabriele@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
beautifulsoup4 = "^4.12.2"
dacite = "^1.8.1"
ipython = "^8.16.0"
jsonc-parser = "^1.1.5"
pyautogen = { git = "https://github.com/microsoft/autogen" }
python = "^3.10"
requests = "^2.31.0"
termcolor = "^2.3.0"
overrides = "^7.4.0"

[tool.poetry.group.dev.dependencies]
pylint = "^3.0.1"
pytest = "^7.4.2"
pytest-mock = "^3.11.1"
types-beautifulsoup4 = "^4.12.0.6"
types-requests = "^2.31.0.7"
mypy = "^1.6.1"

[build-system]
build-backend = "poetry.core.masonry.api"
requires = ["poetry-core"]
